♻️:: 임의값 지정 및 필요없는 코드 삭제
