+++ lib/presentation/minigame/screen/minigame_screen.dart
@@ -5,11 +5,13 @@ import 'package:gogo_app/design_system/component/button/gogo_icon_button.dart';
 import 'package:gogo_app/design_system/theme/color.dart';
 import 'package:gogo_app/design_system/theme/icon.dart';
 import 'package:gogo_app/design_system/theme/typography.dart';
+import 'package:gogo_app/presentation/logIn/screen/login_screen.dart';
 import 'package:gogo_app/presentation/minigame/widgets/minigame_play_component.dart';
 
 class MinigameScreen extends StatelessWidget {
+  /* =========================== 삭제 해야할값
   final int shellgameTickets; // 야바위 티켓수
-  final int cointTossTickets; // 코인토스 티켓수
+  final int cointTossTickets; // 코인토스 티켓수 
   final int plinkoTickets; // 플링코 티켓수
   final int shellgameTicketsCount; // 야바위 티켓 구매 가능한 수수
   final int cointTossTicketsCount; // 코인토스 티켓 구매 가능한 수
@@ -18,18 +20,9 @@ class MinigameScreen extends StatelessWidget {
   final int cointTossTicketscost; // 플링코 티켓 가격
   final int plinkoTicketscost; // 야바위 티켓 가격
   final int point; // 보유 포인트
+  =========================== 삭제 해야할값 */
 
   const MinigameScreen({
-    required this.shellgameTickets,
-    required this.cointTossTickets,
-    required this.plinkoTickets,
-    required this.shellgameTicketsCount,
-    required this.cointTossTicketsCount,
-    required this.plinkoTicketsCount,
-    required this.shellgameTicketscost,
-    required this.cointTossTicketscost,
-    required this.plinkoTicketscost,
-    required this.point,
     super.key,
   });
 
@@ -68,7 +61,7 @@ class MinigameScreen extends StatelessWidget {
                           color: GogoColors.white,
                         ),
                         Text(
-                          '$shellgameTickets',
+                          '0',
                           style: GogoTypography.body3Semibold.copyWith(
                             color: GogoColors.white,
                           ),
@@ -86,7 +79,7 @@ class MinigameScreen extends StatelessWidget {
                           color: GogoColors.white,
                         ),
                         Text(
-                          '$cointTossTickets',
+                          '2',
                           style: GogoTypography.body3Semibold.copyWith(
                             color: GogoColors.white,
                           ),
@@ -104,7 +97,7 @@ class MinigameScreen extends StatelessWidget {
                           color: GogoColors.white,
                         ),
                         Text(
-                          '$shellgameTickets',
+                          '0',
                           style: GogoTypography.body3Semibold.copyWith(
                             color: GogoColors.white,
                           ),
@@ -136,7 +129,7 @@ class MinigameScreen extends StatelessWidget {
                       spacing: 8,
                       children: [
                         Text(
-                          '$point',
+                          '2000',
                           style: GogoTypography.caption1Semibold.copyWith(
                             color: GogoColors.white,
                           ),
@@ -151,156 +144,15 @@ class MinigameScreen extends StatelessWidget {
                   )
                 ],
               ),
-              shellgameTicketscost: shellgameTicketscost,
-              cointTossTicketscost: cointTossTicketscost,
-              plinkoTicketscost: plinkoTicketscost,
-              shellgameTicketsCount: shellgameTicketsCount,
-              cointTossTicketsCount: cointTossTicketsCount,
-              plinkoTicketsCount: plinkoTicketsCount,
+              shellgameTicketscost: 1000,
+              cointTossTicketscost: 1000,
+              plinkoTicketscost: 1000,
+              shellgameTicketsCount: 5,
+              cointTossTicketsCount: 5,
+              plinkoTicketsCount: 5,
             ),
           ],
         ),
-      ),
-    );
-  }
-}
-
-// 미니게임 정보 공통 컴포넌트
-
-class MinigameSelectComponent extends StatelessWidget {
-  final double width;
-  final double height;
-  final Widget gameIcon;
-  final String gameName;
-  final int ticketsCost;
-  final int ticketsCount;
-  final bool info; // 티켓 정보이면 true, 미니게임이면 false,
-
-  const MinigameSelectComponent({
-    super.key,
-    this.width = 343,
-    this.height = 143,
-    required this.gameIcon,
-    required this.gameName,
-    required this.ticketsCost,
-    this.ticketsCount = 0,
-    required this.info,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(spacing: 16.h, children: [
-      Container(
-        padding: EdgeInsets.symmetric(vertical: 14, horizontal: 12),
-        width: width,
-        height: height,
-        decoration: BoxDecoration(
-            color: GogoColors.gray700, borderRadius: BorderRadius.circular(12)),
-        child: Stack(
-          children: [
-            Center(
-              child: Column(
-                spacing: 16,
-                children: [
-                  gameIcon,
-                  Text(
-                    gameName,
-                    style: GogoTypography.body1Semibold.copyWith(
-                      color: GogoColors.white,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            Positioned(
-                right: 0,
-                child: GogoIcons.questionMarkCircle(
-                  onTap: () {},
-                  color: GogoColors.gray500,
-                )),
-          ],
-        ),
-      ),
-      Column(
-        spacing: 8,
-        children: [
-          Container(
-              width: width,
-              height: 45,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(12),
-                color: GogoColors.main600,
-              ),
-              child: TextButton(
-                onPressed: () {},
-                child: Text(
-                  info ? "게임하기" : '$ticketsCost',
-                  style: GogoTypography.caption1Semibold.copyWith(
-                    color: GogoColors.white,
-                  ),
-                ),
-              )),
-          info
-              ? Container()
-              : Text(
-                  '구매 가능한 티켓 : $ticketsCount',
-                  style: GogoTypography.caption2Semibold.copyWith(
-                    color: GogoColors.gray500,
-                  ),
-                ),
-        ],
-      ),
-      SizedBox(
-        height: 12,
-      ),
-    ]);
-  }
-}
-
-// 미니게임 공통 탑바
-
-class MinigameTopBar extends StatelessWidget {
-  final double width;
-  final double height;
-  final Widget icon;
-  final String text;
-  final Widget component;
-
-  const MinigameTopBar({
-    this.width = 343,
-    required this.height,
-    required this.icon,
-    required this.text,
-    required this.component,
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: width.w,
-      height: height,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          SizedBox(
-            width: 67.w,
-            height: height,
-            child: Row(
-              spacing: 8.w,
-              children: [
-                icon,
-                Text(
-                  text,
-                  style: GogoTypography.body2Extrabold.copyWith(
-                    color: GogoColors.white,
-                  ),
-                ),
-              ],
-            ),
-          ),
-          component,
-        ],
       ),
     );
   }
@@ -398,32 +250,6 @@ class MinigameComponent extends StatelessWidget {
           gameName: "플링코",
           ticketsCost: plinkoTicketscost,
           ticketsCount: plinkoTicketsCount,
-        ),
-      ],
-    );
-  }
-}
-
-// 미니게임 탑바 티켓
-
-class MinigameTicket extends StatelessWidget {
-  final int tickets;
-  final Widget gameIcon;
-
-  const MinigameTicket({
-    required this.tickets,
-    required this.gameIcon,
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        gameIcon,
-        Text(
-          '$tickets',
-          style: GogoTypography.body3Semibold,
         ),
       ],
     );
